requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
jinja2==3.1.3
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
groq
